refactor: merge reflective command map lookups into a single getKnownCommands method
