+++ pom.xml
@@ -11,7 +11,7 @@
   </parent>
 
   <artifactId>AlternativeCommandListener</artifactId>
-  <version>3.1-SNAPSHOT</version>
+  <version>3.1</version>
   <packaging>jar</packaging>
 
   <name>AlternativeCommandListener</name>
